requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.7
clipspy==1.0.4
flask==3.0.3
importlib-metadata==8.4.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
networkx==3.2.1
numpy==2.0.2
packaging==24.1
pycparser==2.22
scikit-fuzzy==0.5.0
scipy==1.13.1
werkzeug==3.0.4
zipp==3.20.1
